room, mob, item 以 json 資料為 template, 資料庫只儲存變動資料

+++ src/main/java/com/example/htmlmud/domain/model/RoomStateRecord.java
@@ -2,6 +2,6 @@ package com.example.htmlmud.domain.model;
 
 import java.util.List;
 
-public record RoomStateRecord(int id, List<GameItem> droppedItems) {
+public record RoomStateRecord(long id, List<GameItem> droppedItems) {
 
 }
